# refactor: use more stdlib & optimize dataset/points.Size