--- a/pkg/proxy/tls/options/options.go
+++ b/pkg/proxy/tls/options/options.go
@@ -19,8 +19,6 @@ package options
 import (
 	"os"
 	"slices"
-
-	strutil "github.com/trickstercache/trickster/v2/pkg/util/strings"
 )
 
 // Options is a collection of TLS-related client and server configurations
@@ -71,7 +69,7 @@ func (o *Options) Equal(o2 *Options) bool {
 	return o.FullChainCertPath == o2.FullChainCertPath &&
 		o.PrivateKeyPath == o2.PrivateKeyPath &&
 		o.InsecureSkipVerify == o2.InsecureSkipVerify &&
-		strutil.Equal(o.CertificateAuthorityPaths, o2.CertificateAuthorityPaths) &&
+		slices.Equal(o.CertificateAuthorityPaths, o2.CertificateAuthorityPaths) &&
 		o.ClientCertPath == o2.ClientCertPath &&
 		o.ClientKeyPath == o2.ClientKeyPath
 }