--- a/pkg/util/strings/strings_test.go
+++ b/pkg/util/strings/strings_test.go
@@ -34,22 +34,6 @@ func TestSubstring(t *testing.T) {
 	}
 }
 
-func TestIndexOfString(t *testing.T) {
-
-	arr := []string{"string0", "string1", "string2"}
-
-	i := IndexInSlice(arr, "string0")
-	if i != 0 {
-		t.Errorf(`expected 0. got %d`, i)
-	}
-
-	i = IndexInSlice(arr, "string3")
-	if i != -1 {
-		t.Errorf(`expected -1. got %d`, i)
-	}
-
-}
-
 func TestEqual(t *testing.T) {
 
 	l1 := []string{"test1", "test2"}
@@ -74,6 +58,14 @@ func TestEqual(t *testing.T) {
 	}
 }
 
+func BenchmarkEqual(b *testing.B) {
+	l1 := []string{"test1", "test2"}
+	l2 := []string{"test1", "test2"}
+	for i := 0; i < b.N; i++ {
+		Equal(l1, l2)
+	}
+}
+
 func TestStringMap(t *testing.T) {
 
 	sm := StringMap(map[string]string{"test": "value"})
@@ -82,29 +74,6 @@ func TestStringMap(t *testing.T) {
 
 	if s != expected {
 		t.Errorf("expected %s got %s", expected, s)
-	}
-
-}
-
-func TestCloneBoolMap(t *testing.T) {
-
-	m1 := CloneBoolMap(nil)
-	if m1 != nil {
-		t.Error("expected nil map")
-	}
-
-	const expected = true
-
-	m := map[string]bool{"test": expected}
-	m2 := CloneBoolMap(m)
-
-	v, ok := m2["test"]
-	if !ok {
-		t.Errorf("expected true got %t", ok)
-	}
-
-	if v != expected {
-		t.Errorf("expected %t got %t", expected, v)
 	}
 
 }
@@ -120,6 +89,13 @@ func TestUnique(t *testing.T) {
 	empty := Unique(nil)
 	if len(empty) != 0 {
 		t.Error("expected empty list")
+	}
+}
+
+func BenchmarkUnique(b *testing.B) {
+	initial := []string{"test", "test", "test1", "test2", "test2", "test", "test3"}
+	for i := 0; i < b.N; i++ {
+		Unique(initial)
 	}
 }