* Removes unused code
* Replaces custom string/slice functions with stdlib
* Removes extra `slices.reverse` call in `pkg/router/lm`
* Optimize point size calculation

---
Equal -- ~no change
Unique --  20% improvement

```
BenchmarkEqual-12     	245202459	         4.393 ns/op
BenchmarkUnique-12    	14429709	        88.49 ns/op
```

vs (main):

```
BenchmarkEqual-12     	263193535	         4.449 ns/op
BenchmarkUnique-12    	10249306	       112.2 ns/op
```
---
Points.Size -- looks like a big improvement -- I'm not sure if the goroutine was helping here

```
BenchmarkPointsSize-12    	     100	       654.5 ns/op
```

vs (main):

```
BenchmarkPointsSize-12    	     100	    291119 ns/op
```